Add KPrototypes.fit_predict2() and KPrototypes.predict2().
These functions behave the same as their fit_predict() and predict() versions, but they
return a second list, which is the distance/cost vector with the cost for each point
(distance to the cluster centroid).  The same routine is used for computing the distance
during the fitting as during the predict operation, so the distance should be consistent.